TTK: refactored VTK to Mesh processor
